requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
aioboto3==12.3.0
python-multipart==0.0.9
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
boto3-stubs[s3,dynamodb,ssm,sts]
mangum